requirements.txt
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-mock>=3.11.1
pyyaml>=6.0
